+++ CHANGELOG.md
@@ -4,6 +4,14 @@ This file tracks each version tag to record the additions and changes made.
 
 Using this file will indicate which version to include on the source reference element.
 
+## v1.0.4 - 2025-08-22
+
+Add key-vault module.
+
+## v1.0.3 - 2025-08-22
+
+Update storage-account module variable validation and readme file.
+
 ## v1.0.1 - 2025-07-29
 
 Add storage-account module.